backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sqlalchemy>=2.0.36
pydantic>=2.10.0
pydantic-settings>=2.7.0
python-dotenv>=1.0.1
litellm>=1.55.0
httpx>=0.23.0
beautifulsoup4>=4.12.3
fpdf2>=2.8.0
jinja2>=3.1.5
python-multipart>=0.0.20
aiofiles>=24.1.0
pdfplumber>=0.11.0
python-docx>=1.1.0
